+++ tensorrt_llm/_torch/flashinfer_utils.py
@@ -1,0 +1,27 @@
+import os
+import platform
+import traceback
+
+from ..logger import logger
+
+IS_FLASHINFER_AVAILABLE = False
+
+
+def get_env_enable_pdl():
+    return os.environ.get("TRTLLM_ENABLE_PDL", "0") == "1"
+
+
+ENABLE_PDL = get_env_enable_pdl()
+if ENABLE_PDL:
+    logger.info("PDL is enabled")
+
+if platform.system() != "Windows":
+    try:
+        import flashinfer
+        logger.info(f"flashinfer is available: {flashinfer.__version__}")
+        IS_FLASHINFER_AVAILABLE = True
+    except ImportError:
+        traceback.print_exc()
+        print(
+            "flashinfer is not installed properly, please try pip install or building from source codes"
+        )
